bring no delete org ahead of time

+++ ui/RouteHandlers/OrgRouteHandler.class.php
@@ -981,7 +981,7 @@ class OrgRouteHandler
 
             if (isset($post['deleteId'])) {
                 $deleteId = $post['deleteId'];
-                if ($deleteId) {
+                if (false && $deleteId) {
                     if (!empty($post['sesskey']) && $post['sesskey'] === $sesskey && $orgDao->deleteOrg($org->getId())) {
                         UserRouteHandler::flash(
                             "success",
